requirements.txt
Faker==25.0.1
mysql-connector-python==8.4.0
python-dateutil==2.9.0.post0
six==1.16.0
